Feat : like dislike 구현

+++ findFounder/src/main/java/com/example/demo/service/PostService.java
@@ -16,11 +16,14 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.example.demo.model.dto.PostDto;
 import com.example.demo.model.dto.req.PostCreateReq;
+import com.example.demo.model.dto.req.ReactionReq;
 import com.example.demo.model.dto.res.CommonRes;
 import com.example.demo.model.entity.Customer;
 import com.example.demo.model.entity.Post;
+import com.example.demo.model.entity.Reaction;
 import com.example.demo.repository.CustomerRepository;
 import com.example.demo.repository.PostRepository;
+import com.example.demo.repository.ReactionRepository;
 import com.example.demo.util.JwtUtil;
 
 import lombok.extern.slf4j.Slf4j;
@@ -30,11 +33,14 @@ import lombok.extern.slf4j.Slf4j;
 public class PostService {
 	private PostRepository postRepo;
 	private CustomerRepository customerRepository;
+	private ReactionRepository reactionRepository;
 	
 	@Autowired
-	public PostService(PostRepository postRepository, CustomerRepository customerRepository) {
+	public PostService(PostRepository postRepository, CustomerRepository customerRepository
+			, ReactionRepository reactionRepository) {
 		this.postRepo = postRepository;
 		this.customerRepository = customerRepository;
+		this.reactionRepository = reactionRepository;
 	}
 	
 	// 2024-03-31 CRUD 생성
@@ -62,13 +68,27 @@ public class PostService {
 	
 	// 게시글 수정
 	@Transactional
-	public CommonRes updatePost(int postid, PostCreateReq postDto) {
-		Post post = postRepo.findById(postid).orElseThrow(()->
-		new IllegalArgumentException("해당 게시글이 없습니다."));
-		post.setPostTitle(postDto.getPostTitle());
-		post.setPostContent(postDto.getPostContent());
-		postRepo.saveAndFlush(post);
-		CommonRes commonRes = CommonRes.builder().code(200).msg("게시글 수정 완료").build();
+	public CommonRes updatePost(int cuscode, int postid, PostCreateReq postDto) {
+		
+		Optional<Customer> customerOptional = customerRepository.findById(cuscode);
+		CommonRes commonRes;
+		if (customerOptional.isPresent()) {
+			Post post = getPost(postid);
+			if (cuscode == post.getCustomer().getCusCode()) {
+				post.setPostTitle(postDto.getPostTitle());
+				post.setPostContent(postDto.getPostContent());
+				postRepo.saveAndFlush(post);
+				commonRes = CommonRes.builder().code(200).msg("게시글 수정 완료되었습니다.").build();
+			}
+			else {
+				commonRes = CommonRes.builder().code(300).msg("본인이 아니기 때문에 게시글 수정이 불가능합니다.").build();
+			}	
+		}
+		else {
+			commonRes = CommonRes.builder().code(400).msg("회원 정보가 존재하지 않습니다.").build();
+		}
+			
+	
 		return commonRes;
 	}
 	
@@ -108,10 +128,25 @@ public class PostService {
 	
 	// 게시글 삭제
 	@Transactional
-	public CommonRes deletePost(int post_id) {
+	public CommonRes deletePost(int cuscode, int post_id) {
 		
-		postRepo.deleteById(post_id);
-		CommonRes commonRes = CommonRes.builder().code(200).msg("게시글 삭제 완료").build();
+		Optional<Customer> customerOptional = customerRepository.findById(cuscode);
+		CommonRes commonRes;
+		if (customerOptional.isPresent()) {
+			Post post = getPost(post_id);
+			
+			if (post.getCustomer().getCusCode() == cuscode) {
+				postRepo.deleteById(post_id);
+				commonRes = CommonRes.builder().code(200).msg("게시글 삭제가 완료되었습니다.").build();
+			}
+			else {
+				commonRes = CommonRes.builder().code(300).msg("본인이 아니기 때문에 게시글 삭제가 불가능합니다.").build();
+			}
+		}
+		else {
+			commonRes = CommonRes.builder().code(400).msg("회원 정보가 존재하지 않습니다.").build();
+		}
+		
 		return commonRes;
 	}
 	
@@ -122,5 +157,72 @@ public class PostService {
 	
 	
 	
-	
+	// Like, DisLike
+	@Transactional
+	public CommonRes doReaction(ReactionReq req, int cusCode, int postId) {
+		// reactionId(PK), postId(FK), reaction(1 like, 0 dislike), cusCode(FK)
+		Post post =  postRepo.findById(postId).orElseThrow(() -> new NoSuchElementException("해당하는 게시글이 없습니다."));
+		Customer customer = customerRepository.findById(cusCode).orElseThrow(() -> new NoSuchElementException("해당하는 사용자가 없습니다."));
+		Reaction reaction =   reactionRepository.findByPost_PostIdAndCustomer_CusCode(postId, cusCode);
+		CommonRes commonRes = CommonRes.builder().code(100).msg("초기화").build();;
+		
+		// 1번 케이스 초기의 경우(무조건 생성)
+		if (reaction == null) {
+			reaction = Reaction.builder()
+					.post(post)
+					.customer(customer)
+					.reaction(req.isReaction())
+					.build();
+			reactionRepository.save(reaction);
+			// True인 경우 좋아요
+			if (req.isReaction()) {
+				post.setPostLike(post.getPostLike()+1);
+			}
+			else {
+				post.setPostDislike(post.getPostDislike() + 1);
+			}
+//			postRepo.save(post);
+			commonRes = CommonRes.builder().code(200).msg("리액션 반영 완료").build();
+		}
+		// 이미 True(좋아요)
+		else if (reaction.isReaction()){
+			// 좋아요 2번은 삭제
+			if (req.isReaction()) {
+				reactionRepository.delete(reaction);
+				post.setPostLike(post.getPostLike()-1);
+				
+				commonRes = CommonRes.builder().code(200).msg("리액션 좋아요 2번 클릭으로 인한 삭제 완료").build();
+			}
+			// 좋아요 -> 싫어요
+			else {
+				post.setPostLike(post.getPostLike()-1);
+				post.setPostDislike(post.getPostDislike() + 1);
+				
+				reaction.setReaction(req.isReaction());
+				reactionRepository.save(reaction);
+				commonRes = CommonRes.builder().code(200).msg("리액션 좋아요 후 싫어요 반영 완료").build();
+			}
+		}
+		// 이미 False(싫어요)
+		else if (! reaction.isReaction()) {
+			if (req.isReaction()) {
+				reaction.setReaction(req.isReaction());
+				reactionRepository.save(reaction);
+				
+				post.setPostLike(post.getPostLike() + 1);
+				post.setPostDislike(post.getPostDislike() - 1);
+				
+				commonRes = CommonRes.builder().code(200).msg("리액션 싫어요 후 좋아요 반영 완료").build();
+			}
+			// 
+			else {
+				reactionRepository.delete(reaction);
+				post.setPostDislike(post.getPostDislike() - 1);
+				commonRes = CommonRes.builder().code(200).msg("리액션 싫어요 2번 클릭으로 인한 삭제 완료").build();
+			}
+		}
+		postRepo.save(post);
+		
+		return commonRes;
+	}
 }
